Employer route added sponsor

+++ canopus-backend/routes/employer.router.js
@@ -49,6 +49,13 @@ router.post("/", (req, res) => {
             saved:0,
             closed:0,
         },
+        locumtier:{
+            allowed:1,
+            posted:0,
+            saved:0,
+            closed:0,
+        },
+        sponsors:1,
         validated: false,
     });
     Employer.register(employer, req.body.password)
@@ -228,18 +235,21 @@ router.put("/profile/update/", middleware.isEmployer, (req, res) => {
 //Job related stuff starts here
 //post a job
 router.post("/post/job", middleware.isEmployer, (req, res) => {
-    const expiry=new Date(req.body.endDate);
+    const expiry=new Date(req.body.expireAt);
     var days=(expiry-Date.now())/(1000*60*60*24);
-    if(days<0 || days>90 )res.status(400).json({err:"Invalid time format"});
+    if(days<0 || days>90 )
+    return res.status(400).json({err:"Invalid time format"});
+    else
     Employer.findById(req.user._id).then((employer) => {
 		//validation
 		if(employer.jobtier.allowed-employer.jobtier.posted<=0)
-		res.status(400).send("Max Jobs Posted");
+		return res.status(400).json({err:"Max Jobs Posted"});
 		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"jobtier.posted":1}}).then((employer2)=>{
 	let job = new Job({
 		title: req.body.title,
 		profession: req.body.profession,
-		specialization: req.body.specialization,
+        specialization: req.body.specialization,
+        superSpecialization:req.body.superSpecialization,
 		description: req.body.description,
 		address: req.body.address,
         createdAt:new Date(),
@@ -261,6 +271,7 @@ router.post("/post/job", middleware.isEmployer, (req, res) => {
 				title: req.body.title,
 				profession: req.body.profession,
 				specialization: req.body.specialization,
+                superSpecialization:req.body.superSpecialization,
 				description: req.body.description,
 				address: req.body.address,
                 createdAt:new Date(),
@@ -330,15 +341,27 @@ router.post("/post/job", middleware.isEmployer, (req, res) => {
 router.post("/post/freelance", middleware.isEmployer, (req, res) => {
     const expiry=new Date(req.body.endDate);
     var days=(expiry-Date.now())/(1000*60*60*24);
-    if(days<0 || days>30 )res.status(400).json({err:"Invalid time format"});
+    if(days<0 || days>30 )
+    return res.status(400).json({err:"Invalid time format"});
+    else
     Employer.findById(req.user._id).then((employer) => { 
+        var update={};
+        if(req.body.category=="Day Job"){
 		if(employer.freelancetier.allowed-employer.freelancetier.posted<=0)
-		res.status(400).json({err:"Max Jobs Posted"});
-		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"freelancetier.posted":1}}).then((employer2) =>{
+        return res.status(400).json({err:"Max Jobs Posted"});
+            update={"freelancetier.posted":1};
+        }
+        else{
+		if(employer.locumtier.allowed-employer.locumtier.posted<=0)
+        return res.status(400).json({err:"Max Jobs Posted"});
+            update={"locumtier.posted":1};
+        }
+		Employer.findByIdAndUpdate(req.user._id,{$inc:update}).then((employer2) =>{
 	let freelance = new Freelance({
 		title: req.body.title,
 		profession: req.body.profession,
 		specialization: req.body.specialization,
+        superSpecialization:req.body.superSpecialization,
 		description: req.body.description,
 		address: req.body.address,
 		startDate:req.body.startDate,
@@ -362,6 +385,7 @@ router.post("/post/freelance", middleware.isEmployer, (req, res) => {
 				title: req.body.title,
 				profession: req.body.profession,
 				specialization: req.body.specialization,
+                superSpecialization:req.body.superSpecialization,
 				description: req.body.description,
 				address: req.body.address,
 				startDate:req.body.startDate,
@@ -426,19 +450,21 @@ router.post("/post/freelance", middleware.isEmployer, (req, res) => {
 
 //save a job
 router.post("/save/job", middleware.isEmployer, (req, res) => {
-    const expiry=new Date(req.body.endDate);
+    const expiry=new Date(req.body.expireAt);
     var days=(expiry-Date.now())/(1000*60*60*24);
-    if(days<0 || days>90 )res.status(400).json({err:"Invalid time format"});
+    if(days<0 || days>90 )
+   return res.status(400).json({err:"Invalid time format"});
 	Employer.findById(req.user._id).then((employer) => {
         //validation
 		if(employer.jobtier.allowed-employer.jobtier.saved<=0)
-		res.status(400).send("Max Jobs Saved");
+		return res.status(400).send("Max Jobs Saved");
 		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"jobtier.saved":1}}).then((employer2)=>{
 	let job = new savedJob({
 		status:"Saved",
 		title: req.body.title,
 		profession: req.body.profession,
 		specialization: req.body.specialization,
+        superSpecialization:req.body.superSpecialization,
 		description: req.body.description,
 		address: req.body.address,
         createdAt:new Date(),
@@ -504,18 +530,29 @@ router.post("/save/job", middleware.isEmployer, (req, res) => {
 
 // Save freelance job
 router.post("/save/freelance", middleware.isEmployer, (req, res) => {
-	Employer.findById(req.user._id).then((employer) => {
-        const expiry=new Date(req.body.endDate);
+    const expiry=new Date(req.body.endDate);
 			var days=(expiry-Date.now())/(1000*60*60*24);
-			if(days<0 || days>30 )res.status(400).json({err:"Invalid time format"});
-		if(employer.freelancetier.allowed-employer.freelancetier.posted<=0)
-		res.status(400).json({err:"Max Jobs Posted"});
-		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"freelancetier.saved":1}}).then((employer2) =>{
+            if(days<0 || days>30 )
+            return res.status(400).json({err:"Invalid time format"});
+    Employer.findById(req.user._id).then((employer) => {
+            var update={};
+            if(req.body.category=="Day Job"){
+            if(employer.freelancetier.allowed-employer.freelancetier.posted<=0)
+            return res.status(400).json({err:"Max Jobs Posted"});
+                update={"freelancetier.saved":1};
+            }
+            if(req.body.category=="Locum"){
+            if(employer.locumtier.allowed-employer.locumtier.posted<=0)
+            return res.status(400).json({err:"Max Jobs Posted"});
+                update={"locumtier.saved":1};
+            }
+		else Employer.findByIdAndUpdate(req.user._id,{$inc:update}).then((employer2) =>{
 	let freelance = new savedFreelance({
 		status:"Saved",
 		title: req.body.title,
 		profession: req.body.profession,
 		specialization: req.body.specialization,
+        superSpecialization:req.body.superSpecialization,
 		description: req.body.description,
 		address: req.body.address,
 		startDate:req.body.startDate,
@@ -573,20 +610,22 @@ router.post("/save/freelance", middleware.isEmployer, (req, res) => {
 //validate a saved job
 router.put("/save/job/activate/:id",middleware.isEmployer, (req,res) =>{
 	Employer.findById(req.user._id).then((employer)=>{
-		if(employer.savedJobs.includes(req.params.id))
+        if(!employer.savedJobs.includes(req.params.id))
+        return res.status(400).json({err:"Wrong job ID"});
 		savedJob.findById(req.params.id).then( (sjob) =>{
 			const expiry=sjob.expireAt;
 			var days=(expiry-Date.now())/(1000*60*60*24);
 			if(days<0 || days>90 )
-            res.status(400).send("Invalid time range spcified");
+            return res.status(400).send("Invalid time range spcified");
             if(employer.jobtier.allowed-employer.jobtier.saved<=0)
-		res.status(400).send("Max Jobs Saved");
+		return res.status(400).send("Max Jobs Saved");
 		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"jobtier.posted":1},$inc:{"jobtier.saved":-1}}).then((employer2)=>{
 				let job = new Job({
 					author:sjob.author,
 					title: sjob.title,
 					profession: sjob.profession,
 					specialization: sjob.specialization,
+                    superSpecialization:sjob.superSpecialization,
 					description: sjob.description,
 					address: sjob.address,
 					createdAt:new Date(),
@@ -625,20 +664,22 @@ router.put("/save/job/activate/:id",middleware.isEmployer, (req,res) =>{
 //validate a saved freelance job
 router.put("/save/freelance/activate/:id",middleware.isEmployer, (req,res) =>{
 	Employer.findById(req.user._id).then((employer)=>{
-		if(employer.savedFreelance.includes(req.params.id))
+        if(!employer.savedFreelance.includes(req.params.id))
+        return res.status(400).json({err:"Wrong job ID"});
 		savedFreelance.findById(req.params.id).then( (sjob) =>{
 			const expiry=new Date(sjob.endDate);
 			var days=(expiry-Date.now())/(1000*60*60*24);
 			if(days<0 || days>30 )
-            res.status(400).send("Invalid time range spcified");
+            return res.status(400).send("Invalid time range spcified");
             if(employer.freelancetier.allowed-employer.freelancetier.saved<=0)
-		res.status(400).send("Max Jobs Saved");
-		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"jobtier.posted":1},$inc:{"jobtier.saved":-1}}).then((employer2)=>{
+		return res.status(400).send("Max Jobs Saved");
+		else Employer.findByIdAndUpdate(req.user._id,{$inc:{"freelancetier.posted":1},$inc:{"freelancetier.saved":-1}}).then((employer2)=>{
 				let job = new Freelance({
 					author:sjob.author,
 					title: sjob.title,
 					profession: sjob.profession,
-					specialization: sjob.specialization,
+                    specialization: sjob.specialization,
+                    superSpecialization:sjob.superSpecialization,
 					description: sjob.description,
 					address: sjob.address,
 					createdAt:new Date(),
@@ -679,11 +720,37 @@ router.put("/save/freelance/activate/:id",middleware.isEmployer, (req,res) =>{
 //===========================================================================
 //sponsor a job
 router.put("/sponsor/job/:id",middleware.checkJobOwnership,(req,res)=>{
-	
+	 Employer.findById(req.user._id).then((employer)=>{
+         const Id=employer.jobs.map(item=>{
+             return mongoose.Types.ObjectId(item.id);
+         });
+        if(!Id.includes(req.params.id))
+        return res.status(400).json({err:"Incorrect job ID"});
+        if(employer.sponsors<=0)
+        return res.status(400).json({err:"No sponsors remaining"});
+        Employer.findByIdAndUpdate(req.user._id,{$inc:{sponsors:-1}}).then((employer)=>{
+            Job.findByIdAndUpdate(req.params.id,{$set:{sponsored:"True"}}).then((job)=>{
+                res.json({job:job});
+            }).catch((err)=>{res.status(400).json({err:"Job not updated"})});
+        }).catch((err)=>{res.status(400).json({err:"Employer not found"})});
+     }).catch((err)=>{res.status(400).json({err:"Employer not found"})});
 });
 
 router.put("/sponsor/freelance/:id",middleware.checkFreelanceJobOwnership,(req,res)=>{
-
+    Employer.findById(req.user._id).then((employer)=>{
+        const Id=employer.freelanceJobs.map(item=>{
+            return mongoose.Types.ObjectId(item.id);
+        });
+       if(!Id.includes(req.params.id))
+       return res.status(400).json({err:"Incorrect job ID"});
+       if(employer.sponsors<=0)
+       return res.status(400).json({err:"No sponsors remaining"});
+       Employer.findByIdAndUpdate(req.user._id,{$inc:{sponsors:-1}}).then((employer)=>{
+           Freelance.findByIdAndUpdate(req.params.id,{$set:{sponsored:"True"}}).then((job)=>{
+               res.json({job:job});
+           }).catch((err)=>{res.status(400).json({err:"Job not updated"})});
+       }).catch((err)=>{res.status(400).json({err:"Employer not found"})});
+    }).catch((err)=>{res.status(400).json({err:"Employer not found"})});
 });
 //Update a job
 router.put("/post/job/:id",middleware.checkJobOwnership,async (req,res) =>{
@@ -715,7 +782,8 @@ router.put("/post/job/:id",middleware.checkJobOwnership,async (req,res) =>{
 //Update a freelance job
 router.put("/post/freelance/:id",middleware.checkFreelanceJobOwnership,async (req,res) =>{
     var query;
-    if(req.body.status=="Active") res.status(400).json({err:"Invalid status"});
+    if(req.body.status=="Active") 
+    return res.status(400).json({err:"Invalid status"});
 		query = await searchController.updateQueryBuilder(req)
 	//console.log(query.updateQuery);
 	Freelance.findById(req.params.id).then((job)=>{
@@ -724,7 +792,7 @@ router.put("/post/freelance/:id",middleware.checkFreelanceJobOwnership,async (re
 	const expiry=new Date(req.body.endDate);
 		var days=(expiry-job.createdAt)/(1000*60*60*24) ;
 		if(days<0 || days>90 )
-		res.status(400).json({err:"Invalid time format"});
+		return res.status(400).json({err:"Invalid time format"});
 		else
 		query.update["expireAt"]=expiry;
 	// if(req.body.sponsored) {
@@ -741,19 +809,20 @@ router.put("/post/freelance/:id",middleware.checkFreelanceJobOwnership,async (re
 //Update a saved job
 router.put("/save/job/:id",middleware.isEmployer,async (req,res) =>{
     var query; 
-	if(req.body.status=="Active") res.status(400).json({err:"Invalid status"});
+    if(req.body.status=="Active") 
+    return res.status(400).json({err:"Invalid status"});
 		query = await searchController.updateQueryBuilder(req)
 	//console.log(query.updateQuery);
 	Employer.findById(req.user._id).then((employer)=>{
 	if(!employer.savedJobs.includes(req.params.id))
-	res.status(400).json({err:"Invalid Job"});
+	return res.status(400).json({err:"Invalid Job"});
 	savedJob.findById(req.params.id).then((job)=>{
 	if(req.body.expireAt){
 	const expiry=new Date(req.body.expireAt);
 		var days=(expiry-Date.now())/(1000*60*60*24) ;
 		//console.log(days);
 		if(days<0 || days>90 )
-		res.status(400).json({err:"Invalid time format"});
+		return res.status(400).json({err:"Invalid time format"});
 		else
 		query.update["expireAt"]=expiry;
 	}
@@ -773,19 +842,20 @@ router.put("/save/job/:id",middleware.isEmployer,async (req,res) =>{
 //Update a saved freelance job
 router.put("/save/freelance/:id",middleware.isEmployer,async (req,res) =>{
     var query;
-    if(req.body.status=="Active") res.status(400).json({err:"Invalid status"});
+    if(req.body.status=="Active") 
+    return res.status(400).json({err:"Invalid status"});
 		query = await searchController.updateQueryBuilder(req)
 	//console.log(query.updateQuery);
 	Employer.findById(req.user._id).then((employer)=>{
 	if(!employer.savedFreelance.includes(req.params.id))
-	res.status(400).json({err:"Invalid Job"});
+	return res.status(400).json({err:"Invalid Job"});
 	savedFreelance.findById(req.params.id).then((job)=>{
 	if(req.body.endDate){
 	const expiry=new Date(req.body.endDate);
 		var days=(expiry-Date.now())/(1000*60*60*24) ;
 		//console.log(days);
 		if(days<0 || days>90 )
-		res.status(400).json({err:"Invalid time format"});
+		return res.status(400).json({err:"Invalid time format"});
 		else
 		query.update["expireAt"]=expiry;
 	}
@@ -942,7 +1012,7 @@ Promise.all(promises)
 .catch((err) => res.json({ err:"Couldn't find employer" }));
 });
 //get saved job
-router.get("/savedJob/:id",middleware.isEmployer,(req,res)=>{
+router.get("/save/job/:id",middleware.isEmployer,(req,res)=>{
 savedJob.findById(req.params.id)
 .then((job) =>{
     res.json(job);
@@ -950,7 +1020,7 @@ savedJob.findById(req.params.id)
 .catch((err) =>res.status(400).json({err:"Job not found"}));
 });
 //get saved freelance job
-router.get("/savedFreelance/:id",middleware.isEmployer,(req,res)=>{
+router.get("/save/freelance/:id",middleware.isEmployer,(req,res)=>{
 savedFreelance.findById(req.params.id)
 .then((job) =>{
     res.json(job);
@@ -958,7 +1028,7 @@ savedFreelance.findById(req.params.id)
 .catch((err) =>res.status(400).json({err:"Job not found"}));
 });
 //get a job by id
-router.get("job/:id", (req, res) => {
+router.get("/post/job/:id", (req, res) => {
 Job.findById(req.params.id)
     .then((job) => {
         res.json(job);
@@ -966,7 +1036,7 @@ Job.findById(req.params.id)
     .catch((err) => res.status(400).json({ err: err }));
 });
 //get a freelance job by id
-router.get("/freelance/:id", (req, res) => {
+router.get("/post/freelance/:id", (req, res) => {
 Freelance.findById(req.params.id)
     .then((job) => {
         res.json(job);
@@ -976,7 +1046,7 @@ Freelance.findById(req.params.id)
 
 //delete a job
 
-router.delete("job/:id", middleware.checkJobOwnership, (req, res) => {
+router.delete("/post/job/:id", middleware.checkJobOwnership, (req, res) => {
 Job.findByIdAndDelete(req.params.id)
 .then(() => res.json("Job deleted successfully !"))
 .catch((err) =>
@@ -987,18 +1057,20 @@ Job.findByIdAndDelete(req.params.id)
 });
 //delete a job
 
-router.delete("/savedJob/:id", middleware.isEmployer, (req, res) => {
-savedJob.findByIdAndDelete(req.params.id)
-.then(() => res.json("Saved Job deleted successfully !"))
-.catch((err) =>
-       res.status(400).json({
-           err: err,
-       }),
-       );
+router.delete("/save/job/:id", middleware.isEmployer, (req, res) => {
+    Employer.findById(req.user._id).then((employer)=>{
+        if(employer.savedJobs.includes(req.params.id))
+            savedJob.findByIdAndDelete(req.params.id)
+            .then(() => res.json("Saved Job deleted successfully !"))
+            .catch((err) =>
+                 res.status(400).json({
+                 err: err,
+                 }),);
+    }).catch((err)=>{res.json({err:"Job doesn't belong to you"})});
 });
 //delete a job
 
-router.delete("/freelance/:id", middleware.checkFreelanceJobOwnership, (req, res) => {
+router.delete("/post/freelance/:id", middleware.checkFreelanceJobOwnership, (req, res) => {
 Freelance.findByIdAndDelete(req.params.id)
 .then(() => res.json("Freelance Job deleted successfully !"))
 .catch((err) =>
@@ -1009,15 +1081,19 @@ Freelance.findByIdAndDelete(req.params.id)
 });
 //delete a job
 
-router.delete("/savedFreelance/:id", middleware.isEmployer, (req, res) => {
-savedFreelance.findByIdAndDelete(req.params.id)
+router.delete("/save/freelance/:id", middleware.isEmployer, (req, res) => {
+    Employer.findById(req.user._id).then((employer)=>{
+        if(employer.savedJobs.includes(req.params.id))
+            savedFreelance.findByIdAndDelete(req.params.id)
 .then(() => res.json("Saved Freelance Job deleted successfully !"))
 .catch((err) =>
        res.status(400).json({
            err: err,
        }),
        );
+    }).catch((err)=>{res.json({err:"Job doesn't belong to you"})});
 });
+
 module.exports = router;
 /*
     "username":"sparshjain",
